# docs/api-contract/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "VideoInstructionCreate": {
        "description": "Request schema for creating a video instruction tutorial.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "target_department": {
            "default": "All",
            "title": "Target Department",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "video_url": {
            "title": "Video Url",
            "type": "string"
          }
        },
        "required": [
          "title",
          "video_url"
        ],
        "title": "VideoInstructionCreate",
        "type": "object"
      },
      "VideoInstructionResponse": {
        "description": "Response schema containing video details and total view count.",
        "properties": {
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "is_archived": {
            "default": false,
            "title": "Is Archived",
            "type": "boolean"
          },
          "tags": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tags"
          },
          "target_department": {
            "default": "All",
            "title": "Target Department",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "video_url": {
            "title": "Video Url",
            "type": "string"
          },
          "views_count": {
            "title": "Views Count",
            "type": "integer"
          }
        },
        "required": [
          "title",
          "video_url",
          "id",
          "created_at",
          "views_count"
        ],
        "title": "VideoInstructionResponse",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "title": "Magti Internal Portal API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/videos": {
      "get": {
        "description": "Retrieves video instruction modules.\n\nRegular operators and managers see only videos targeting their department or \"All\".\nContent administrators and system administrators see all videos.\n\nAccess: Authenticated users (any active role).\n\nArgs:\n    current_user: The authenticated User object.\n    db: SQLAlchemy database session.\n\nReturns:\n    A list of VideoInstructionResponse schemas.",
        "operationId": "get_videos_api_videos_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/VideoInstructionResponse"
                  },
                  "title": "Response Get Videos Api Videos Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Get Videos",
        "tags": [
          "videos"
        ]
      },
      "post": {
        "description": "Creates a new video instruction entry, sends real-time SSE notify, and logs the action.\n\nAccess: Restricted to content administrators (content_admin) and system administrators (admin).\n\nArgs:\n    video: Video instruction creation details.\n    current_admin: The authenticated administrator User.\n    db: SQLAlchemy database session.\n\nReturns:\n    The created VideoInstruction database row.",
        "operationId": "create_video_api_videos_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoInstructionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInstructionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Create Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{video_id}": {
      "delete": {
        "description": "Deletes a video instruction entry and logs the action.\n\nAccess: Restricted to content administrators (content_admin) and system administrators (admin).\n\nArgs:\n    video_id: ID of the video to delete.\n    current_admin: The authenticated administrator User.\n    db: SQLAlchemy database session.\n\nReturns:\n    None.\n\nRaises:\n    HTTPException: 404 Not Found if the video does not exist.",
        "operationId": "delete_video_api_videos__video_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Delete Video",
        "tags": [
          "videos"
        ]
      },
      "put": {
        "description": "Updates an existing video instruction entry and logs the action.\n\nAccess: Restricted to content administrators (content_admin) and system administrators (admin).\n\nArgs:\n    video_id: ID of the video to update.\n    video: Video update details.\n    current_admin: The authenticated administrator User.\n    db: SQLAlchemy database session.\n\nReturns:\n    The updated VideoInstruction database row.\n\nRaises:\n    HTTPException: 404 Not Found if the video does not exist.",
        "operationId": "update_video_api_videos__video_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VideoInstructionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInstructionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Update Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{video_id}/archive": {
      "post": {
        "description": "Soft-archives a video instruction by setting is_archived=True.\n\nMirrors archive_article: archived videos disappear from the operator-facing\nlist (get_videos / global search already filter on is_archived for\nnon-admins). Admins continue to see them.",
        "operationId": "archive_video_api_videos__video_id__archive_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInstructionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Archive Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{video_id}/unarchive": {
      "post": {
        "description": "Restores an archived video to is_archived=False.",
        "operationId": "unarchive_video_api_videos__video_id__unarchive_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInstructionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Unarchive Video",
        "tags": [
          "videos"
        ]
      }
    },
    "/api/videos/{video_id}/view": {
      "post": {
        "description": "Increments the view count of a video instruction module.\n\nAccess: Authenticated users (any active role).\n\nArgs:\n    video_id: ID of the video to view.\n    current_user: The authenticated User object.\n    db: SQLAlchemy database session.\n\nReturns:\n    The updated VideoInstruction database row.\n\nRaises:\n    HTTPException: 404 Not Found if the video does not exist.",
        "operationId": "view_video_api_videos__video_id__view_post",
        "parameters": [
          {
            "in": "path",
            "name": "video_id",
            "required": true,
            "schema": {
              "title": "Video Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VideoInstructionResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "View Video",
        "tags": [
          "videos"
        ]
      }
    },
...
  }
}